add Solid Earth projection
there was also a bug where the Liquid Earth projection threw an error if you tried to use a map with points on the pole or antimeridian.  but that's fixd now.  I also fixed a minor issue where it said that the Equal Earth projection had a closed-form inverse when it in fact does not.  I also renamed the liquid earth mesh files.