fix spiral map for good
well, there are a couple of nubs I don't understand that appear when you turn the number of loops down.  but this is fine, I think.  it finally hits the good enuff metric for which I had been aiming.  I just had to pull out the poles and handle them totally separately.