faster root-finding
I implemented a form of the "false position" method, which I've previusly used in Dunier.  I'm not sure why it's called that – I didn't know it had a name until today – but it's about 2× faster than the basic bisection algorithm I was using before, at least for the chinese projection.  I'm curius about this so-called "Ridder's method", which is supposed to be even better at root-finding, but it's late and I've been working on this for too long already, and it works well enuff.