even faster root-finding
it turns out the main reason my method of false-position implementation was so slow was my termination condition – checking the width of the x interval – which just doesn't make sense if you're not bisecting.  I tried Ridders's algorithm as well but it wasn't much faster than what I have so I didn't think it was worth complicating the code so much.