Change 等差分纬线多圆锥 to 1980 implementation
I was using an approximation of this map from 2012, but I was able to find some much older ones from circa 1980.  they're still only fits (I was hoping for original equations) but going closer to the source seems better.  especially since it kind of seems like most subsequent authors just copied existing fits.  it's not clear to me whether anything actually should have changed here, but there were some discrepancies in the tables that seem like they may have been typos (either by me or the subsequent authors), and it was pointed out to me that these all assume the obliqity of the earth is 23.5° (which it's _not_!  it's closer to 23.44° right now!  if they were going to use an outated approximation they should have just said "23.5" instead of saying "tropic of cancer"!).  also, can I just say, the units in both of the books that were provided to me are abhorrent.  one uses radians for latitude and degrees for longitude, while the other uses the reverse.  why???